+++ crates/key-manager/src/lib.rs
@@ -1,8 +1,0 @@
-#![feature(try_blocks)]
-
-mod error;
-mod key_manager;
-mod persistence;
-
-pub use error::KeyManagerError;
-pub use key_manager::KeyManager;