+++ crates/server-config/src/resolved_config.rs
@@ -46,9 +46,12 @@ pub struct UnresolvedConfig {
 
 impl UnresolvedConfig {
     pub fn resolve(self) -> eyre::Result<ResolvedConfig> {
+        let node_config = self.node_config.resolve(&self.dir_config.base_dir)?;
+        let dir_config = self.dir_config.resolve()?;
+
         Ok(ResolvedConfig {
-            dir_config: self.dir_config.resolve()?,
-            node_config: self.node_config.resolve()?,
+            dir_config,
+            node_config,
         })
     }
 }