+++ crates/workers/src/error.rs
@@ -19,11 +19,17 @@ use std::path::PathBuf;
 use thiserror::Error;
 
 #[derive(Debug, Error)]
-pub enum KeyManagerError {
+pub enum KeyStorageError {
     #[error("Failed to persist keypair: RSA is not supported")]
     CannotExtractRSASecretKey,
     #[error("Error reading persisted keypair from {path:?}: {err}")]
     ReadPersistedKeypair {
+        path: PathBuf,
+        #[source]
+        err: std::io::Error,
+    },
+    #[error("Error reading list of directory from {path:?}: {err}")]
+    DirectoryListError {
         path: PathBuf,
         #[source]
         err: std::io::Error,
@@ -64,6 +70,53 @@ pub enum KeyManagerError {
         #[source]
         err: std::io::Error,
     },
+    #[error("Error creating directory for persisted keypairs {path:?}: {err}")]
+    CreateKeypairsDir {
+        path: PathBuf,
+        #[source]
+        err: std::io::Error,
+    },
+
+    #[error("Keypair for peer_id {0} not found")]
+    KeypairNotFound(PeerId),
+}
+
+#[derive(Debug, Error)]
+pub enum WorkersError {
+    #[error("Error creating directory for persisted workers {path:?}: {err}")]
+    CreateWorkersDir {
+        path: PathBuf,
+        #[source]
+        err: std::io::Error,
+    },
+    #[error("Error creating key pair for worker: {err}")]
+    CreateWorkerKeyPair {
+        #[source]
+        err: KeyStorageError,
+    },
+    #[error("Error removing key pair for worker: {err}")]
+    RemoveWorkerKeyPair {
+        #[source]
+        err: KeyStorageError,
+    },
+    #[error("Error reading persisted worker from {path:?}: {err}")]
+    ReadPersistedWorker {
+        path: PathBuf,
+        #[source]
+        err: std::io::Error,
+    },
+    #[error("Error deserializing persisted worker from {path:?}: {err}")]
+    DeserializePersistedWorker {
+        path: PathBuf,
+        #[source]
+        err: toml::de::Error,
+    },
+    #[error("Worker for {deal_id} already exists")]
+    WorkerAlreadyExists { deal_id: String },
+    #[error("Worker for deal_id {0} not found")]
+    WorkerNotFoundByDeal(String),
+    #[error("Worker {0} not found")]
+    WorkerNotFound(PeerId),
     #[error("Error serializing persisted worker: {err}")]
     SerializePersistedWorker {
         #[source]
@@ -82,24 +135,6 @@ pub enum KeyManagerError {
         #[source]
         err: std::io::Error,
     },
-    #[error("Error creating directory for persisted keypairs {path:?}: {err}")]
-    CreateKeypairsDir {
-        path: PathBuf,
-        #[source]
-        err: std::io::Error,
-    },
-    #[error("Error creating directory for persisted workers {path:?}: {err}")]
-    CreateWorkersDir {
-        path: PathBuf,
-        #[source]
-        err: std::io::Error,
-    },
     #[error("Keypair for peer_id {0} not found")]
     KeypairNotFound(PeerId),
-    #[error("Worker for {deal_id} already exists")]
-    WorkerAlreadyExists { deal_id: String },
-    #[error("Worker for deal_id {0} not found")]
-    WorkerNotFoundByDeal(String),
-    #[error("Worker {0} not found")]
-    WorkerNotFound(PeerId),
 }