+++ sorcerer/src/worker_builins.rs
@@ -18,37 +18,39 @@ use fluence_spell_dtos::trigger_config::TriggerConfig;
 use futures::TryFutureExt;
 use serde_json::Value as JValue;
 use std::str::FromStr;
+use std::sync::Arc;
 use std::time::Duration;
 
 use crate::spell_builtins::remove_spell;
-use key_manager::KeyManager;
 use particle_args::{Args, JError};
 use particle_execution::ParticleParams;
 use particle_services::ParticleAppServices;
 use spell_event_bus::api::{from_user_config, SpellEventBusApi};
 use spell_service_api::{CallParams, SpellServiceApi};
 use spell_storage::SpellStorage;
+use workers::{Scope, Workers};
 
-pub(crate) fn create_worker(
+pub(crate) async fn create_worker(
     args: Args,
     params: ParticleParams,
-    key_manager: KeyManager,
+    workers: Arc<Workers>,
 ) -> Result<JValue, JError> {
     let mut args = args.function_args.into_iter();
     let deal_id: String = Args::next("deal_id", &mut args)?;
     Ok(JValue::String(
-        key_manager
-            .create_worker(deal_id, params.init_peer_id)?
+        workers
+            .create_worker(deal_id, params.init_peer_id)
+            .await?
             .to_base58(),
     ))
 }
 
-pub(crate) fn get_worker_peer_id(args: Args, key_manager: KeyManager) -> Result<JValue, JError> {
+pub(crate) fn get_worker_peer_id(args: Args, workers: Arc<Workers>) -> Result<JValue, JError> {
     let mut args = args.function_args.into_iter();
     let deal_id: String = Args::next("deal_id", &mut args)?;
 
     Ok(JValue::Array(
-        key_manager
+        workers
             .get_worker_id(deal_id)
             .map(|id| vec![JValue::String(id.to_base58())])
             .unwrap_or_default(),
@@ -58,7 +60,7 @@ pub(crate) fn get_worker_peer_id(args: Args, key_manager: KeyManager) -> Result<
 pub(crate) async fn remove_worker(
     args: Args,
     params: ParticleParams,
-    key_manager: KeyManager,
+    workers: Arc<Workers>,
     services: ParticleAppServices,
     spell_storage: SpellStorage,
     spell_event_bus_api: SpellEventBusApi,
@@ -66,13 +68,13 @@ pub(crate) async fn remove_worker(
     let mut args = args.function_args.into_iter();
     let worker_id: String = Args::next("worker_id", &mut args)?;
     let worker_id = PeerId::from_str(&worker_id)?;
-    let worker_creator = key_manager.get_worker_creator(worker_id)?;
+    let worker_creator = workers.get_worker_creator(worker_id)?;
 
     if params.init_peer_id != worker_creator && params.init_peer_id != worker_id {
         return Err(JError::new(format!("Worker {worker_id} can be removed only by worker creator {worker_creator} or worker itself")));
     }
 
-    tokio::task::spawn_blocking(move || key_manager.remove_worker(worker_id)).await??;
+    workers.remove_worker(worker_id).await?;
 
     let spells: Vec<_> = spell_storage.get_registered_spells_by(worker_id);
 
@@ -97,9 +99,9 @@ pub(crate) async fn remove_worker(
     Ok(())
 }
 
-pub(crate) fn worker_list(key_manager: KeyManager) -> Result<JValue, JError> {
+pub(crate) fn worker_list(workers: Arc<Workers>) -> Result<JValue, JError> {
     Ok(JValue::Array(
-        key_manager
+        workers
             .list_workers()
             .into_iter()
             .map(|p| JValue::String(p.to_base58()))
@@ -110,7 +112,8 @@ pub(crate) fn worker_list(key_manager: KeyManager) -> Result<JValue, JError> {
 pub(crate) async fn deactivate_deal(
     args: Args,
     params: ParticleParams,
-    key_manager: KeyManager,
+    workers: Arc<Workers>,
+    scope: Scope,
     spell_storage: SpellStorage,
     spell_event_bus_api: SpellEventBusApi,
     spell_service_api: SpellServiceApi,
@@ -118,16 +121,15 @@ pub(crate) async fn deactivate_deal(
     let mut args = args.function_args.into_iter();
     let deal_id: String = Args::next("deal_id", &mut args)?;
 
-    if !key_manager.is_management(params.init_peer_id) && !key_manager.is_host(params.init_peer_id)
-    {
+    if !scope.is_management(params.init_peer_id) && !scope.is_host(params.init_peer_id) {
         return Err(JError::new(format!(
             "Only management or host peer can deactivate deal"
         )));
     }
 
-    let worker_id = key_manager.get_worker_id(deal_id)?;
+    let worker_id = workers.get_worker_id(deal_id)?;
 
-    if !key_manager.is_worker_active(worker_id) {
+    if !workers.is_worker_active(worker_id) {
         return Err(JError::new("Deal has already been deactivated"));
     }
 
@@ -159,7 +161,7 @@ pub(crate) async fn deactivate_deal(
             })?;
     }
 
-    tokio::task::spawn_blocking(move || key_manager.deactivate_worker(worker_id)).await??;
+    workers.deactivate_worker(worker_id).await?;
 
     Ok(())
 }
@@ -167,7 +169,8 @@ pub(crate) async fn deactivate_deal(
 pub(crate) async fn activate_deal(
     args: Args,
     params: ParticleParams,
-    key_manager: KeyManager,
+    workers: Arc<Workers>,
+    scope: Scope,
     services: ParticleAppServices,
     spell_event_bus_api: SpellEventBusApi,
     spell_service_api: SpellServiceApi,
@@ -176,16 +179,15 @@ pub(crate) async fn activate_deal(
     let mut args = args.function_args.into_iter();
     let deal_id: String = Args::next("deal_id", &mut args)?;
 
-    if !key_manager.is_management(params.init_peer_id) && !key_manager.is_host(params.init_peer_id)
-    {
+    if !scope.is_management(params.init_peer_id) && !scope.is_host(params.init_peer_id) {
         return Err(JError::new(format!(
             "Only management or host peer can activate deal"
         )));
     }
 
-    let worker_id = key_manager.get_worker_id(deal_id)?;
+    let worker_id = workers.get_worker_id(deal_id)?;
 
-    if key_manager.is_worker_active(worker_id) {
+    if workers.is_worker_active(worker_id) {
         return Err(JError::new("Deal has already been activated"));
     }
 
@@ -219,13 +221,13 @@ pub(crate) async fn activate_deal(
         })
         .await?;
 
-    tokio::task::spawn_blocking(move || key_manager.activate_worker(worker_id)).await??;
+    workers.activate_worker(worker_id).await?;
     Ok(())
 }
 
-pub(crate) fn is_deal_active(args: Args, key_manager: KeyManager) -> Result<JValue, JError> {
+pub(crate) fn is_deal_active(args: Args, workers: Arc<Workers>) -> Result<JValue, JError> {
     let mut args = args.function_args.into_iter();
     let deal_id: String = Args::next("deal_id", &mut args)?;
-    let worker_id = key_manager.get_worker_id(deal_id)?;
-    Ok(JValue::Bool(key_manager.is_worker_active(worker_id)))
+    let worker_id = workers.get_worker_id(deal_id)?;
+    Ok(JValue::Bool(workers.is_worker_active(worker_id)))
 }